Compute InlineInfo eagerly again to not need Context in opt/
