+++ compiler/src/dotty/tools/backend/jvm/BTypes.scala
@@ -6,8 +6,10 @@ import scala.language.unsafeNulls
 import java.util.concurrent.ConcurrentHashMap
 import scala.tools.asm
 import dotty.tools.backend.jvm.BTypes.InternalName
+import dotty.tools.backend.jvm.opt.OptimizerWarning
 import dotty.tools.dotc.core.Symbols.ClassSymbol
 
+import scala.collection.SortedMap
 import scala.tools.asm.Opcodes
 import scala.tools.asm.tree.{ClassNode, ModuleNode}
 
@@ -555,12 +557,6 @@ sealed trait RefBType extends BType {
  * TODO: innerclass attributes on mirror class
  */
 
-/** Source of a ClassInfo, used to load further information such as inlining-related info */
-enum ClassInfoSource:
-  case Symbol(classSym: ClassSymbol, internalName: InternalName)
-  case Classfile(classNode: ClassNode, moduleNode: Option[ModuleNode])
-  case Missing
-
 /**
  * The type info for a class. Used for symboltable-independent subtype checks in the backend.
  *
@@ -575,11 +571,11 @@ enum ClassInfoSource:
  *                      InnerClass table, see the InnerClass spec summary above.
  *
  * @param nestedInfo    If this describes a nested class, information for the InnerClass table.
- * @param source        Information about how this class was sourced.
+ * @param inlineInfo    Inlining information for this class, if requested and available.
  */
 final case class ClassInfo(superClass: Option[ClassBType], interfaces: List[ClassBType], flags: Int,
                            nestedClasses: List[ClassBType], nestedInfo: Option[NestedInfo],
-                           source: ClassInfoSource)
+                           inlineInfo: InlineInfo)
 
 /**
  * Information required to add a class to an InnerClass table.
@@ -615,6 +611,55 @@ case class NestedInfo(enclosingClass: ClassBType,
  * @param flags     The flags for this class in the InnerClass entry.
  */
 case class InnerClassEntry(name: String, outerName: String, innerName: String, flags: Int)
+
+
+/**
+ * Metadata about a ClassBType, used by the inliner.
+ *
+ * More information may be added in the future to enable more elaborate inline heuristics.
+ * Note that this class should contain information that can only be obtained from the ClassSymbol.
+ * Information that can be computed from the ClassNode should be added to the call graph instead.
+ *
+ * @param isEffectivelyFinal True if the class cannot have subclasses: final classes, module
+ *                           classes.
+ *
+ * @param sam                If this class is a SAM type, the SAM's "\$name\$descriptor".
+ * @param methodInfos        The [[MethodInlineInfo]]s for the methods declared in this class.
+ *                           The map is indexed by the string s"\$name\$descriptor" (to
+ *                           disambiguate overloads).
+ *
+ * @param warning            Contains a warning message if an error occurred when building this
+ *                           InlineInfo, for example if some classfile could not be found on
+ *                           the classpath. This warning can be reported later by the inliner.
+ *
+ * @param isAccessible       Whether this class's internals can be inlined into callsites, i.e.,
+ *                           it is exported from a public module.
+ */
+final case class InlineInfo(isEffectivelyFinal: Boolean,
+                            sam: Option[String],
+                            methodInfos: collection.SortedMap[(String, String), MethodInlineInfo],
+                            warning: Option[OptimizerWarning],
+                            isAccessible: Boolean)
+
+object InlineInfo {
+  val empty = InlineInfo(isEffectivelyFinal = false, sam = None, methodInfos = SortedMap.empty, warning = None, isAccessible = false)
+}
+
+trait InlineInfoLoader {
+  def loadInlineInfoFor(name: InternalName): InlineInfo
+}
+
+/**
+ * Metadata about a method, used by the inliner.
+ *
+ * @param effectivelyFinal  True if the method cannot be overridden (in Scala)
+ * @param annotatedInline   True if the method is annotated `@inline`
+ * @param annotatedNoInline True if the method is annotated `@noinline`
+ */
+final case class MethodInlineInfo(effectivelyFinal: Boolean = false,
+                                  annotatedInline: Boolean = false,
+                                  annotatedNoInline: Boolean = false)
+
 
 /**
  * A ClassBType represents a class or interface type.
